Extract Components::ApplicationForm::FieldWrapperRendering
TextField and TextareaField had byte-for-byte identical show_label?,
inline?, wrapper_class, and form_group_class methods, plus a
near-identical render_with_wrapper (label -> [prepend] -> field
content -> help -> append). #4842 added a :prepend slot to
TextareaField only, since that PR's scope didn't touch TextField --
flagged there as a DRY gap that would recur the next time a text
field needed the same slot.

Extracts the shared skeleton into a new module, sibling to
FieldLabelRow/FieldWithHelp. prepend/append presence is guarded via
respond_to?, the same pattern FieldLabelRow already uses for
label_end_slot, so field types that don't declare a slot just no-op
rather than needing their own override.

No behavior change -- confirmed against the existing component/form
test suite covering both field types.
