+++ app/components/application_form/text_field.rb
@@ -6,6 +6,7 @@ class Components::ApplicationForm < Superform::Rails::Form
     include Phlex::Slotable
     include FieldWithHelp
     include FieldLabelRow
+    include FieldWrapperRendering
     include InputGroupAddon
 
     slot :between
@@ -15,8 +16,6 @@ class Components::ApplicationForm < Superform::Rails::Form
 
     # Make slot accessors public (Phlex::Slotable makes them private by default)
     public :between_slot, :label_end_slot, :append_slot, :help_slot
-
-    attr_reader :wrapper_options
 
     def initialize(field, wrapper_options: {}, **attributes)
       super(field, **attributes)
@@ -29,38 +28,19 @@ class Components::ApplicationForm < Superform::Rails::Form
                                                "form-control"))
       else
         render_with_wrapper do
-          input(**attributes, class: class_names(attributes[:class],
-                                                 "form-control"))
+          render_field_input do
+            input(**attributes, class: class_names(attributes[:class],
+                                                   "form-control"))
+          end
         end
       end
     end
 
     private
 
-    def render_with_wrapper(&field_input)
-      div(class: wrapper_class, data: wrapper_options[:wrap_data]) do
-        render_label_row(label_text, inline?) if show_label?
-        render_field_input(&field_input)
-        render_help_after_field
-        render(append_slot) if append_slot
-      end
-    end
-
     def bare_input?
       attributes[:type] == "hidden" ||
         (wrapper_options[:label] == false && !help_slot)
-    end
-
-    def show_label?
-      wrapper_options[:label] != false
-    end
-
-    def inline?
-      wrapper_options[:inline] || false
-    end
-
-    def wrapper_class
-      form_group_class("form-group", inline?, wrapper_options[:wrap_class])
     end
 
     def render_field_input(&block)
@@ -71,13 +51,6 @@ class Components::ApplicationForm < Superform::Rails::Form
       else
         yield
       end
-    end
-
-    def form_group_class(base, inline, wrap_class)
-      classes = base
-      classes += " form-inline" if inline && base == "form-group"
-      classes += " #{wrap_class}" if wrap_class.present?
-      classes
     end
   end
 end
